dataflow_pipeline/requirements.txt
google-cloud-pubsub==2.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*
google-cloud-dataflow==0.8.*